Acabadas las vistas de delete.php de todas las entidades. Se ha añadido un input hidden para volver a pasar el ID del elemento a eliminar, ya que no funcionaba cuando se confirmaba el borrado.

+++ views/directors/delete.php
@@ -7,23 +7,48 @@
         <div class="container">
             <?php
                 $idDirector = $_POST['directorId'];
-                $directorDeleted = deleteDirector($idDirector);
+                $deleteConfirmed = false;
 
-                if ($directorDeleted) {
+                if (isset($_POST['confirmBtn'])) {
+                    $deleteConfirmed = true;
+                }
+
+                if ($deleteConfirmed) {
+                    $directorDeleted = deleteDirector($idDirector);
+                }
+                
+                if (!$deleteConfirmed) {
             ?>
-                    <div class="row">
-                        <div class="alert alert-success" role="alert">
-                            Director borrado correctamente.<br><a href="list.php">Volver al listado de directores.</a>
-                        </div>
+                    <div class="alert alert-warning" role="alert">
+                        ¿Estás seguro de que quieres eliminar el registro?
+                    </div>
+                    <div class="btn-group" role="group" aria-label="Basic example">
+                        <form name="confirm_delete" action="" method="POST">
+                            <input type="submit" value="Confirmar" class="btn btn-success" name="confirmBtn"/>
+                            <input type="hidden" name="directorId" value="<?php echo $idDirector; ?>"/>
+                        </form>                      
+                        <a class="btn btn-danger" href="list.php">Cancelar</a>                                           
                     </div>
             <?php
                 } else {
+                    if ($directorDeleted) {
             ?>
-                    <div class="row">
-                        <div class="alert alert-danger" role="alert">
-                            El director no se ha borrado correctamente. Asegurese de que no pertenezca a ninguna serie.<br><a href="list.php">Volver a intentarlo.</a>
+                        <div class="row">
+                            <div class="alert alert-success" role="alert">
+                                Director borrado correctamente.<br><a href="list.php">Volver al listado de directores.</a>
+                            </div>
                         </div>
-                    </div>
+            <?php
+                    } else {
+            ?>
+                        <div class="row">
+                            <div class="alert alert-danger" role="alert">
+                                El director no se ha borrado correctamente. Asegurese de que no pertenezca a ninguna serie.<br><a href="list.php">Volver a intentarlo.</a>
+                            </div>
+                        </div>
+            <?php
+                    }
+            ?>
             <?php
                 }
             ?>
